fixed OfficeScene
