+++ client/src/App.js
@@ -1,4 +1,3 @@
-import EditorComponent from "./components/pages/editor";
 import Routes from "./components/routes";
 function App() {
   return (
