+++ client/src/components/pages/editor/Editor.js
@@ -3,6 +3,7 @@ import * as dat from 'dat.gui';
 import ToolBox from './tools/ToolBox';
 import ObjectGenerator from './viewport/utils/ObjectGenerator';
 import CameraSelector from './viewport/CameraSelector';
+import { Vector3 } from 'three';
 
 export default class Editor{
     constructor(viewportCanvas, toolBarElement, propertiesPaneContainer){
@@ -55,8 +56,10 @@ export default class Editor{
         this.sceneOutliner.open();
         this.objectGenerator = new ObjectGenerator(this.viewport, this.sceneOutliner, this.cameraSelector);
         this.bindAddOption();
-        this.objectGenerator.addCube();
-        
+    
+        //add initial objects
+        this.initObjects();
+
         //add render option
         this.renderMode = false;
         this.propertiesPane.add(this, 'renderMode').name('Render').onChange(()=>{
@@ -64,6 +67,14 @@ export default class Editor{
             if(this.renderMode)
                 this.propertiesPane.close();
         });
+    }
+
+    initObjects(){
+        this.objectGenerator.addCube();
+        this.objectGenerator.addAmbientLight();
+        this.objectGenerator.cursorPoint = new Vector3(-4,3,2);
+        this.objectGenerator.addDirectionalLight();
+        this.objectGenerator.cursorPoint = new Vector3(0,0,0);
     }
 
     bindCameraProperties(){
@@ -105,8 +116,17 @@ export default class Editor{
         addMeshFolder.add(this.objectGenerator, 'addCone').name('Cone');
         addMeshFolder.add(this.objectGenerator, 'addTorus').name('Torus');
         addMeshFolder.add(this.objectGenerator, 'addText').name('Text');
-        this.loadHelicopter = ()=>{
-            this.objectGenerator.addObj('./assets/editor/models/Seahawk.obj', 'Helicopter');
+        this.loadHelicopter = () => {
+            this.objectGenerator.addObj(
+                './assets/editor/models/Seahawk.obj',
+                true,
+                'Helicopter',
+                (helicopter) => {
+                    helicopter.scale.x *= 0.1;
+                    helicopter.scale.y *= 0.1;
+                    helicopter.scale.z *= 0.1;
+                }
+            );
         };
         addOptionFolder.add(this, 'loadHelicopter').name('Helicopter');
         addOptionFolder.add(this.objectGenerator, 'addCamera').name('Camera');
