+++ client/src/components/pages/editor/viewport/utils/ObjectGenerator.js
@@ -23,6 +23,8 @@ import RectAreaLightProperty from "../propertyController/lightPropertyController
 import SpotLightProperty from "../propertyController/lightPropertyController/SpotLightProperty";
 import PointLightProperty from "../propertyController/lightPropertyController/PointLightProperty";
 import InteractiveModel from "../interactiveObjects/InteractiveModel";
+import {FontLoader} from 'three/examples/jsm/loaders/FontLoader';
+import {TextGeometry} from 'three/examples/jsm/geometries/TextGeometry';
 
 export default class ObjectGenerator {
     static OBJECT_TYPE = {
@@ -57,7 +59,7 @@ export default class ObjectGenerator {
 
         this.loadingManager = new THREE.LoadingManager();
         this.objLoader = new OBJLoader(this.loadingManager);
-        this.fontLoader = new THREE.FontLoader(this.loadingManager);
+        this.fontLoader = new FontLoader(this.loadingManager);
         
     }
 
@@ -69,7 +71,7 @@ export default class ObjectGenerator {
 
     createPlane(attachProperties = true) {
         let properties;
-        let geometry = new  THREE.PlaneGeometry(1, 1);
+        let geometry = new THREE.PlaneGeometry(1, 1);
         let material = new THREE.MeshStandardMaterial({ color: 0x8e9091, side: THREE.DoubleSide });
         let mesh = new InteractiveMesh(this.viewport, geometry, material);
         mesh.position.copy(this.cursorPoint);
@@ -90,7 +92,7 @@ export default class ObjectGenerator {
 
     createCube(attachProperties = true) {
         let properties;
-        let geometry = new  THREE.BoxGeometry(1, 1, 1);
+        let geometry = new THREE.BoxGeometry(1, 1, 1);
         let material = new THREE.MeshStandardMaterial({ color: 0x8e9091 });
         let mesh = new InteractiveMesh(this.viewport, geometry, material);
         mesh.position.copy(this.cursorPoint);
@@ -111,7 +113,7 @@ export default class ObjectGenerator {
 
     createCircle(attachProperties = true) {
         let properties;
-        let geometry = new  THREE.CircleGeometry(1, 10);
+        let geometry = new THREE.CircleGeometry(1, 10);
         let material = new THREE.MeshStandardMaterial({ color: 0x8e9091 });
         let mesh = new InteractiveMesh(this.viewport, geometry, material);
         mesh.position.copy(this.cursorPoint);
@@ -132,7 +134,7 @@ export default class ObjectGenerator {
 
     createUVSphere(attachProperties = true) {
         let properties;
-        let geometry = new  THREE.SphereGeometry(1, 30, 30);
+        let geometry = new THREE.SphereGeometry(1, 30, 30);
         let material = new THREE.MeshStandardMaterial({ color: 0x8e9091 });
         let mesh = new InteractiveMesh(this.viewport, geometry, material);
         mesh.position.copy(this.cursorPoint);
@@ -153,7 +155,7 @@ export default class ObjectGenerator {
 
     createIcoSphere(attachProperties = true) {
         let properties;
-        let geometry = new  THREE.IcosahedronGeometry(1, 2);
+        let geometry = new THREE.IcosahedronGeometry(1, 2);
         let material = new THREE.MeshStandardMaterial({ color: 0x8e9091 });
         let mesh = new InteractiveMesh(this.viewport, geometry, material);
         mesh.position.copy(this.cursorPoint);
@@ -174,7 +176,7 @@ export default class ObjectGenerator {
 
     createCylinder(attachProperties = true) {
         let properties;
-        let geometry = new  THREE.CylinderGeometry(1, 1, 1, 20, 20);
+        let geometry = new THREE.CylinderGeometry(1, 1, 1, 20, 20);
         let material = new THREE.MeshStandardMaterial({ color: 0x8e9091 });
         let mesh = new InteractiveMesh(this.viewport, geometry, material);
         mesh.position.copy(this.cursorPoint);
@@ -195,7 +197,7 @@ export default class ObjectGenerator {
 
     createCone(attachProperties = true) {
         let properties;
-        let geometry = new  THREE.ConeGeometry(1, 2, 10, 10);
+        let geometry = new THREE.ConeGeometry(1, 2, 10, 10);
         let material = new THREE.MeshStandardMaterial({ color: 0x8e9091 });
         let mesh = new InteractiveMesh(this.viewport, geometry, material);
         mesh.position.copy(this.cursorPoint);
@@ -216,7 +218,7 @@ export default class ObjectGenerator {
 
     createTorus(attachProperties = true) {
         let properties;
-        let geometry = new  THREE.TorusGeometry(.5, 0.1, 10, 50);
+        let geometry = new THREE.TorusGeometry(.5, 0.1, 10, 50);
         let material = new THREE.MeshStandardMaterial({ color: 0x8e9091 });
         let mesh = new InteractiveMesh(this.viewport, geometry, material);
         mesh.position.copy(this.cursorPoint);
@@ -230,7 +232,7 @@ export default class ObjectGenerator {
     }
 
     createText(text='graphicsAI', font, attachProperties = true){
-        let geometry = new  THREE.TextGeometry(text, {
+        let geometry = new TextGeometry(text, {
             font,
             size: 0.5,
             height: 0.2,
@@ -304,7 +306,7 @@ export default class ObjectGenerator {
     }
 
     createLight(lightHelper, LightPropertyController, attachProperties=true){
-        // lightHelper.light.position.copy(this.cursorPoint);
+        lightHelper.light.position.copy(this.cursorPoint);
         let interactiveLight = Object.assign(lightHelper, new InteractiveLight(this.viewport, lightHelper));
         if(attachProperties){
             lightHelper.properties = new LightPropertyController(interactiveLight, this.propertiesPane);
@@ -394,7 +396,7 @@ export default class ObjectGenerator {
         this.viewport.add(object);
     }
 
-    addObj(objFile, attachProperties = true, name='Obj'){
+    addObj(objFile, attachProperties = true, name='Obj', onAfterAdd){
         this.objLoader.load(
             objFile,
             (object)=>{
@@ -402,6 +404,7 @@ export default class ObjectGenerator {
                     this.attachPropertiesToObj(object, name);
                 }
                 this.viewport.add(object);
+                onAfterAdd(object);
             },
             (xhr)=>{
                 console.log( ( xhr.loaded / xhr.total * 100 ) + '% loaded' );
